Project 1: Space-invaders game
This game is made using python's pygame library and it's a game where you can shoot alien spaceships with your spaceship using bullets until they reach a specific region. It has a cool background and sounds added to it. This was a great experience!! 
It took me a week to do this project.